SDMMCv1: fix system hang issues
1. SDMMCv1: unlock before waiting for DMA completion

Avoids system hanging forever when an error interrupt is signaled and
the controller is in a state where the DMA can never complete.

2. SDMMCv1: wait for DMA completion only when it will complete

In certain error states (e.g. CRC error in the middle of a multi-block
transmission), the DMA may never complete. Therefore, only wait for its
completion if the transfer was successful. We do need to wait for
completion on read as some words may still be in flight from the receive
FIFO to RAM.

The STM32F7 reference manual (RM0385 Rev 9 Section 35.3.2 SDMMC APB2
interface) mentions that it is possible for a receive FIFO overflow on
the last word to be signaled only after the DATAEND flag is set.
Therefore we also have to check that the FIFO does not have data in it
and confirm there is no RXOVERR error too.

3. SDMMCv1: listen for reserved start bit error

Sometimes this is the only interrupt that gets signaled in certain error
cases. We need to listen for it so we don't sleep forever waiting for an
interrupt from the controller.

The STM32F7 reference manual RM0385 documents this bit as reserved but
evidently it can still be triggered in some cases. The logic has been
copied from the SDIOv1 driver where it is not reserved.