SDIOv1: wait for DMA completion only when it will complete
In certain error states (e.g. CRC error in the middle of a multi-block
transmission), the DMA may never complete. Therefore, only wait for its
completion if the transfer was successful. We do need to wait for
completion on read as some words may still be in flight from the receive
FIFO to RAM.