SDMMCv2: reset peripheral on stop to clear state machine
The reference manual sanctioned way to clear the data path state machine
(DPSM) is to use the command path state machine (CPSM) to send a command
with the `CMD` register `CMDSTOP` bit set.

The driver doesn't use that bit so fix it in a brute force way by just
resetting the peripheral on shutdown. Our application already shuts down
the driver on error before trying to restart it.

This avoids the peripheral being unable to work again and the driver
sleeping forever waiting for an interrupt because the DPSM is stuck and
won't accept new transfers during card init.

A better patch might change the driver to properly use the `CMDSTOP`
bit, and send a stop command on startup.